refactor: move the vault opitons into the vault package
